==> teamwork/2차과제/제출/team2/01. world_bank.ipynb <==
import requests
import pandas as pd
import numpy as np
import json

def world_bank_api(target_country, indicat):  
    api_params = {
        'date' : '1960:2021',
        'format' : 'json',
        'frequency' : 'Y',
        'per_page': 20000
    }  
    resp = requests.get(f'http://api.worldbank.org/v2/country/{target_country}/indicator/{indicat}', params=api_params)
    return json.loads(resp.text)

indicator = [
    'NY.GDP.MKTP.CD',       # GDP
    'SP.POP.TOTL',          # 인구
    'EN.ATM.PM25.MC.M3',    # 초미세먼지
    'EN.ATM.CO2E.PC'        # 이산화탄소 배출
    ]
country = 'all'

world_gdp = world_bank_api(country, indicator[0])
world_pop = world_bank_api(country, indicator[1])
world_pm25 = world_bank_api(country, indicator[2])
world_co2e = world_bank_api(country, indicator[3])

df_gdp = pd.json_normalize(world_gdp[1])
df_pop = pd.json_normalize(world_pop[1])
df_pm25 = pd.json_normalize(world_pm25[1])
df_co2e = pd.json_normalize(world_co2e[1])

import pickle

df_gdp['date'] = pd.to_numeric(df_gdp['date'])
df_pop['date'] = pd.to_numeric(df_pop['date'])
df_pm25['date'] = pd.to_numeric(df_pm25['date'])
df_co2e['date'] = pd.to_numeric(df_co2e['date'])

dfs = [
    df_gdp,
    df_pop,
    df_pm25,
    df_co2e
]

for df in dfs:
    df.drop(columns=['unit', 'obs_status', 'decimal', 'indicator.id'], inplace=True)
    df.rename(columns={'indicator.value':'measure', 'country.id':'country_code', 'country.value': 'country'}, inplace=True)

df_gdp.rename(columns={'value': 'gdp'}, inplace=True)
df_pop.rename(columns={'value': 'population'}, inplace=True)
df_pm25.rename(columns={'value': 'pm25'}, inplace=True)
df_co2e.rename(columns={'value': 'co2e'}, inplace=True)

df_co2e.columns

for df in dfs:
    df.drop(columns=['measure'], inplace=True)

ttl_df = pd.merge(df_gdp, df_pop, how='inner', on=['countryiso3code', 'date', 'country_code','country'])

ttl_df = pd.merge(ttl_df, df_co2e, how='inner', on=['countryiso3code', 'date', 'country_code','country'])
ttl_df = pd.merge(ttl_df, df_pm25, how='inner', on=['countryiso3code', 'date', 'country_code','country'])

with open('data/world_bank_all.pkl', 'wb') as f:
    pickle.dump(ttl_df, f)

with open('data/world_bank_gdp.pkl', 'wb') as f:
    pickle.dump(df_gdp, f)

with open('data/world_bank_pop.pkl', 'wb') as f:
    pickle.dump(df_pop, f)

with open('data/world_bank_pm25.pkl', 'wb') as f:
    pickle.dump(df_pm25, f)

with open('data/world_bank_co2e.pkl', 'wb') as f:
    pickle.dump(df_co2e, f)

df_gdp[df_gdp['value'] == df_gdp['value'].max()]

df_pm25.groupby('country')['value'].max()

df_co2e[df_co2e['country']=='China'].to_csv('china_co2e.csv')

df_co2e['value'].max(), df_co2e['value'].min()

df_pm25['value'].max(), df_pm25['value'].min()

df_pop['value'].max(), df_pop['value'].min()

df_gdp['value'].max(), df_gdp['value'].min()

max_len = 0
for item in df_gdp['country']:
    # print(item)
    # print(len(item))
    # break
    max_len = max(len(item), max_len)

print(max_len)

df_sl = pd.read_csv('data/2.Sea_Level.csv', index_col=0)
df_clm = pd.read_csv('data/1.climate.csv', index_col=0)

df_sl.rename(columns={'year':'Year'}, inplace=True)

combined_df = pd.merge(df_clm, df_sl, how='inner', on=['Year'])

with open('data/global.pkl', 'wb') as f:
    pickle.dump(combined_df, f)
