==> teamwork/2차과제/제출/team2/01-1.sea_level.ipynb <==
import pandas as pd
import numpy as np
import matplotlib as mpl
import matplotlib.pyplot as plt

sea_level = pd.read_csv('data/Global_sea_level_rise.csv', encoding='utf-8')


sea_level.rename(columns={"mmfrom1993-2008average":"Sea level"}, inplace=True)

sea_level2 = pd.DataFrame(sea_level)

sea_level2.columns

del sea_level2['date']

sea_level3 = sea_level2[sea_level2['year'].between(1960, 2020)]


sea_level4 = sea_level3.reset_index(drop=True)


sea_level4.to_csv('data/Sea_Level.csv', encoding='utf-8', index=True)

sea_level4.info()

sea_level4["Sea level"].max(), sea_level4["Sea level"].min()
