==> teamwork/2차과제/제출/team2/01-2.climate_data.ipynb <==
import pandas as pd
import numpy as np
import matplotlib as mpl
import matplotlib.pyplot as plt

climate = pd.read_csv("data/Climate Data.csv", encoding='utf-8')

climate2 = pd.DataFrame(climate)

climate3 = climate2.drop(climate2.columns[5:20], axis=1)


climate4 = climate3.drop(climate3.columns[11:15], axis=1)

climate5 = climate4.drop(climate4.columns[1], axis=1)

climate6 = climate5.groupby('Year').mean().reset_index()


climate7 =climate6.drop(0).reset_index(drop=True)


climate8 = climate7.rename(columns={"Land.6":"NoPol_Land","Ocean.6":"NoPol_Ocean","Ocean.7":"SoPol_Ocean","Land.7":"SoPol_Land"}, inplace=False)

climate8.to_csv('data/1.climate.csv', encoding='utf-8', index=True)

climate8.info()

climate8.iloc[:][1:].max(), climate8.iloc[:][1:].min()
