==> teamwork/2차과제/제출/team2/03. get_data.ipynb <==
import cx_Oracle as cx
import oracledb
import pickle
import pandas as pd
import numpy as np


oracledb.init_oracle_client(lib_dir="/Users/minjoongkim/instantclient_19_8")

conn = cx.connect(user="lab02", password="lab02", dsn="localhost:1521/xe")
cursor = conn.cursor()
cursor.execute("select * from country_data")
country_data = cursor.fetchall()
country_df = pd.DataFrame(country_data, columns=[column[0] for column in cursor.description])
cursor.execute("select * from global_data")
global_data = cursor.fetchall()
global_df = pd.DataFrame(global_data, columns=[column[0] for column in cursor.description])
conn.close()

merged_df = pd.merge(country_df, global_df, how='inner', on='year')

merged_df.info()

regions = {
    "Asia" : {
        "Korea"             :       "KOR",
        "China"             :       "CHN",
        "India"             :       "IND",
        "Thailand"          :       "THA"
    },
    "Africa" : {
        "South Africa"      :       "ZAF",
        "Kenya"             :       "KEN",
        "Egypt"             :       "EGY",
        "Morocco"           :       "MAR",
        "Nigeria"           :       "NGA"
    },
    "Europe" : {
        "Russia"            :       "RUS",
        "Germany"           :       "DEU",
        "United Kingdom"    :       "GBR",
        "Spain"             :       "ESP",
        "Italy"             :       "ITA",
        "Norway"            :       "NOR"
    },
    "America" : {
        "United States"     :       "USA",
        "Canada"            :       "CAN",
        "Mexico"            :       "MEX",
        "Brazil"            :       "BRA",
        "Argentina"         :       "ARG",
    },
    "Middle East" : {
        "Turkey"            :       "TUR",
        "Iran"              :       "IRN",
        "Saudi Arabia"      :       "SAU"
    },
    "Oceania" : {
        "Austrailia"        :       "AUS"
    }
}

merged_df['country_name'].value_counts()

pd.DataFrame(merged_df['country_name'].unique()).to_csv("Country_names.csv")

merged_df.columns

new_regions = [
    "Arab World",
    "East Asia & Pacific",
    "European Union",
    "High income",
    "Latin America & Caribbean",
    "Low income",
    "Middle income",
    "North America",
    "OECD members",
    "South Asia",
    "Sub-Saharan Africa",
    "World"
]

for region in new_regions:
    new_df = merged_df[merged_df['country_name'] == region]
    with open(f"data/regional_data/{'_'.join(region.split(' '))}.pkl", "wb") as f:
        pickle.dump(new_df, f)
